requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.16.16
botocore==1.19.16
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-bootstrap4==2.3.1
django-filter==2.4.0
django-storages==1.10.1
django-tinymce4-lite==1.8.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
jsmin==2.2.2
kiwisolver==1.3.1
matplotlib==3.3.3
mpmath==1.1.0
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
sympy==1.6.2
urllib3==1.25.11
virtualenv==20.1.0
whitenoise==5.2.0
